Added Rifle and Shotgun prefab and added platform functionality

+++ Assets/Scripts/GunScripts/Gun.cs
@@ -4,7 +4,7 @@ public class Gun : MonoBehaviour
 {
     public Transform muzzlePoint;
     public int magazineCount;
-    public int knockbackForce;
+    public float knockbackForce;
     public int range;
     public int bulletSpeed;
     public string gunName;
